requirements.txt
blinker==1.8.2
cachetools==5.3.3
click==8.1.7
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
itsdangerous==2.2.0
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
plotly==5.22.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
tenacity==8.3.0
tzdata==2024.1
uritemplate==4.1.1
Werkzeug==3.0.3
